Show errors when database connection test fails (#77)
I've setup the database on a couple different octoprint instances in the last month, and I keep forgetting to install psycopg2. Debugging this is always frustrating as I keep tweaking settings, checking logs and chrome debugger to try to determine the errors, etc. I figured it would make it a lot easier for everyone to include an error message after we got a result from "Test Connection".

This just exposes a bit more info from the responses in the test api, then adds an extra field to the UI to show the returning error. The error clears after a few seconds.